requirements.txt
datadreamer.dev<=0.35.0
openai<=1.55.3
datasets<=2.19.0
transformers<=4.40.1
accelerate<=0.29.3
matplotlib<=3.9.0
seaborn<=0.13.2
playwright<=1.44.0
beautifulsoup4<=4.12.3
PyLaTeX<=1.4.2
pdf2image<=1.17.0
graphviz<=0.20.3
pandas<=2.2.2
scipy<=1.14.0
networkx<=3.3
geopandas<=1.0.0
shapely<=2.0.4
altair<=5.3.0
vl-convert-python<=1.5.0
plotly<=5.22.0
kaleido<=0.2.1.post1
vega_datasets<=0.9.0
altair-saver<=0.5.0
vl-convert-python<=1.5.0
mpl_finance<=0.10.1
mplfinance<=0.12.10b0
weasyprint<=62.3
openpyxl<=3.1.5
dataframe_image<=0.2.4
webdriver_manager<=4.0.1
huggingface_hub<=0.23.5